added new exceptions

+++ tracker/tracker/src/main/java/com/RoomyExpense/tracker/service/ExpenseService.java
@@ -4,6 +4,9 @@ import com.RoomyExpense.tracker.DTO.ExpenseCreationDTO;
 import com.RoomyExpense.tracker.DTO.ExpenseDTO;
 import com.RoomyExpense.tracker.DTO.ExpenseUpdateDTO;
 import com.RoomyExpense.tracker.DTO.HouseDTO;
+import com.RoomyExpense.tracker.exception.ExpenseNotFoundException;
+import com.RoomyExpense.tracker.exception.HouseNotFoundException;
+import com.RoomyExpense.tracker.exception.UserNotFoundException;
 import com.RoomyExpense.tracker.mapper.ExpenseMapper;
 import com.RoomyExpense.tracker.mapper.HouseMapper;
 import com.RoomyExpense.tracker.model.Expense;
@@ -45,18 +48,19 @@ public class ExpenseService implements IExpenseService {
 
     @Override
     public Optional<ExpenseDTO> getExpenseById(Long id){
-        Optional<Expense> expenseOptional = expenseRepository.findById(id);
-        return expenseOptional.map(expenseMapper::toDTO);
+        Expense expense  = expenseRepository.findById(id)
+                .orElseThrow(()-> new ExpenseNotFoundException("Expense with ID " + id + " not found"));
+        return Optional.of(expenseMapper.toDTO(expense));
     }
 
     @Transactional
     @Override
     public Expense createExpense(ExpenseCreationDTO expenseCreationDTO) {
-        // Validate and search the house
+
         Optional<HouseDTO> houseOptional = houseService.getHouseById(expenseCreationDTO.getHouseId());
 
         if (houseOptional.isEmpty()) {
-            throw new EntityNotFoundException("House not found.");
+            throw new HouseNotFoundException("House with ID " + expenseCreationDTO.getHouseId() + " not found");
         }
 
         House house = houseRepository.findById(houseOptional.get().getId())
@@ -75,6 +79,10 @@ public class ExpenseService implements IExpenseService {
     @Transactional
     @Override
     public void deleteExpense(Long id){
+
+        if (!expenseRepository.existsById(id)) {
+            throw new UserNotFoundException("Expense with ID " + id + " not found");
+        }
         expenseRepository.deleteById(id);
     }
 
@@ -83,7 +91,7 @@ public class ExpenseService implements IExpenseService {
     public Expense updateExpense(Long expenseId, ExpenseUpdateDTO expenseUpdateDTO) {
 
         Expense expense = expenseRepository.findById(expenseId)
-                .orElseThrow(() -> new EntityNotFoundException("Expense with ID " + expenseId + " not found"));
+                .orElseThrow(() -> new ExpenseNotFoundException("Expense with ID " + expenseId + " not found"));
 
         if (expenseUpdateDTO.getName() != null) {
             expense.setName(expenseUpdateDTO.getName());
@@ -102,7 +110,7 @@ public class ExpenseService implements IExpenseService {
         }
         if (expense.getHouse() != null) {
             House house = houseRepository.findById(expense.getHouse().getId())
-                    .orElseThrow(() -> new EntityNotFoundException("House with ID " + expense.getHouse().getId() + " not found"));
+                    .orElseThrow(() -> new HouseNotFoundException("House with ID " + expense.getHouse().getId() + " not found"));
             expense.setHouse(house);
         }
